fix(compose): surface network_mode bridge, aux_addresses, restart_policy drops (#522)
## What

Three compose directives that diverged or were dropped without surfacing
to the operator (#511). All now flow through the parse-time diagnostics
pass, the same channel every other unsupported field uses:

1. **`network_mode: bridge` divergence.** docker-compose attaches to
Docker's predefined shared `bridge` network; Podman reads `--network
bridge` as "create a fresh, isolated bridge netns", so the container has
connectivity but cannot reach its project siblings — silently. New
`diverging_network_mode` diagnostic flags this and points at
`networks:`.
2. **`ipam.aux_addresses` visibility.** The drop was announced via a
runtime `tracing::warn!` in `network.rs`, invisible without `RUST_LOG`.
Moved into the diagnostics collector; removed the engine-time warn so it
surfaces once, like the rest.
3. **`deploy.restart_policy.delay` / `.window`.** Parsed but never
consumed (only `condition`/`max_attempts` map to a Podman restart
policy). Now reported instead of dropped silently.

## Tests

- `warns_on_network_mode_bridge_divergence`
- `warns_on_ipam_aux_addresses`
- `warns_on_restart_policy_delay_and_window`

Full lib suite green; fmt clean; clippy clean.

Closes #511
